+++ shared/src/main/scala/com/raquo/domtypes/generic/defs/eventProps/FormEventProps.scala
@@ -3,16 +3,23 @@ package com.raquo.domtypes.generic.defs.eventProps
 import com.raquo.domtypes.generic.builders.EventPropBuilder
 
 /**
+  * @tparam DomHtmlElementTargetEvent
+  *            An event that has an HTMLElement as `target`.
+  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
+  * @tparam DomFormElementTargetEvent
+  *            An event that has an HTMLFormElement as `target`.
+  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
   * @tparam DomInputElementTargetEvent
   *            An event that has an HTMLInputElement as `target`.
-  *            This event type has no corresponding type in JS DOM. See our own `InputElementTargetEvent` trait.
-  *            Strictly speaking, this type is not applicable to `onInput` because the latter
-  *            can be fired on any HTMLElement that has `contentEditable` mode enabled.
+  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
+  *            - Strictly speaking, this type is not applicable to `onInput` because the latter
+  *              can be fired on any HTMLElement that has `contentEditable` mode enabled.
+  *            - Similarly for `onChange` and `onSelect` – these could also fire on an `HTMLTextAreaElement`.
   * @tparam DomInputEvent
   *            DOM InputEvent https://developer.mozilla.org/en-US/docs/Web/API/InputEvent
   *            Note: This type is not implemented in scala-js-dom
   */
-trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, DomInputEvent <: DomEvent, DomElementTargetEvent <: DomEvent] { this: EventPropBuilder[EP, DomEvent] =>
+trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, DomInputEvent <: DomEvent, DomHtmlElementTargetEvent <: DomEvent, DomFormElementTargetEvent <: DomEvent, DomInputElementTargetEvent <: DomEvent] { this: EventPropBuilder[EP, DomEvent] =>
 
   /**
     * The change event is fired for input, select, and textarea elements
@@ -20,7 +27,7 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onChange: EP[DomEvent with DomElementTargetEvent] = eventProp("change")
+  lazy val onChange: EP[DomHtmlElementTargetEvent] = eventProp("change")
 
   /**
     * The select event only fires when text inside a text input or textarea is
@@ -28,20 +35,20 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onSelect: EP[DomEvent with DomElementTargetEvent] = eventProp("select")
+  lazy val onSelect: EP[DomHtmlElementTargetEvent] = eventProp("select")
 
   /**
     * The input event is fired for input, select, textarea, and
     * contentEditable elements when it gets user input.
     */
-  lazy val onInput: EP[DomInputEvent with DomElementTargetEvent] = eventProp("input")
+  lazy val onInput: EP[DomHtmlElementTargetEvent] = eventProp("input")
 
   /**
     * The blur event is raised when an element loses focus.
     *
     * MDN
     */
-  lazy val onBlur: EP[DomFocusEvent with DomElementTargetEvent] = eventProp("blur")
+  lazy val onBlur: EP[DomFocusEvent with DomHtmlElementTargetEvent] = eventProp("blur")
 
 
   /**
@@ -49,7 +56,7 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onFocus: EP[DomFocusEvent with DomElementTargetEvent] = eventProp("focus")
+  lazy val onFocus: EP[DomFocusEvent with DomHtmlElementTargetEvent] = eventProp("focus")
 
   /**
     * The submit event is fired when the user clicks a submit button in a form
@@ -57,22 +64,22 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onSubmit: EP[DomEvent with DomElementTargetEvent] = eventProp("submit")
+  lazy val onSubmit: EP[DomFormElementTargetEvent] = eventProp("submit")
 
   /**
     * The reset event is fired when a form is reset.
     *
     * MDN
     */
-  lazy val onReset: EP[DomEvent with DomElementTargetEvent] = eventProp("reset")
+  lazy val onReset: EP[DomFormElementTargetEvent] = eventProp("reset")
 
   /**
     * Script to be run when an element is invalid
     */
-  lazy val onInvalid: EP[DomEvent with DomElementTargetEvent] = eventProp("invalid")
+  lazy val onInvalid: EP[DomHtmlElementTargetEvent] = eventProp("invalid")
 
   /**
     * Fires when the user writes something in a search field (for <input="search">)
     */
-  lazy val onSearch: EP[DomEvent with DomElementTargetEvent] = eventProp("search")
+  lazy val onSearch: EP[DomInputElementTargetEvent] = eventProp("search")
 }